+++ src/KairosWeb_Groep6/Controllers/Baten/ExtraBesparingenController.cs
@@ -169,7 +169,7 @@ namespace KairosWeb_Groep6.Controllers.Baten
                 ViewData["totaal"] = 0;
             }
 
-            double totaal = analyse.ExtraBesparingen
+            decimal totaal = analyse.ExtraBesparingen
                                     .Sum(t => t.Bedrag);
 
             ViewData["totaal"] = totaal.ToString("C", new CultureInfo("nl-BE"));
