+++ src/KairosWeb_Groep6/Controllers/Kosten/VoorbereidingsKostenController.cs
@@ -168,7 +168,7 @@ namespace KairosWeb_Groep6.Controllers.Kosten
                 ViewData["totaal"] = 0;
             }
 
-            double totaal = KostOfBaatExtensions.GeefTotaal(analyse.VoorbereidingsKosten);
+            decimal totaal = KostOfBaatExtensions.GeefTotaal(analyse.VoorbereidingsKosten);
 
             ViewData["totaal"] = totaal.ToString("C", new CultureInfo("nl-BE"));
         }
